requirements.txt
fastapi
uvicorn [standard]
sqlalchemy [asyncio]
aiosqlite
passlib [bcrypt]
python-jose
pydantic 